web-server/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
autopep8==2.3.2
click==8.2.1
fastapi==0.104.1
h11==0.16.0
httptools==0.6.4
idna==3.10
psycopg2-binary==2.9.9
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
pytz==2023.3
PyYAML==6.0.2
redis==5.0.1
sniffio==1.3.1
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
